Move output to pkg
