+++ pkg/osvscanner/vulnerability_result.go
@@ -3,10 +3,10 @@ package osvscanner
 import (
 	"fmt"
 
-	"github.com/google/osv-scanner/internal/output"
 	"github.com/google/osv-scanner/pkg/grouper"
 	"github.com/google/osv-scanner/pkg/models"
 	"github.com/google/osv-scanner/pkg/osv"
+	"github.com/google/osv-scanner/pkg/output"
 )
 
 // groupResponseBySource converts raw OSV API response into structured vulnerability information
